Remove Validator\TypeTestCase and add validator logic to base TypeTestCase
