+++ src/Symfony/Component/Form/Test/TypeTestCase.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Form\Test;
 
 use Symfony\Component\Form\FormBuilder;
 use Symfony\Component\EventDispatcher\EventDispatcher;
+use Symfony\Component\Form\Test\Traits\ValidatorExtensionTrait;
 
 abstract class TypeTestCase extends FormIntegrationTestCase
 {
@@ -32,6 +33,24 @@ abstract class TypeTestCase extends FormIntegrationTestCase
 
         $this->dispatcher = $this->getMockBuilder('Symfony\Component\EventDispatcher\EventDispatcherInterface')->getMock();
         $this->builder = new FormBuilder(null, null, $this->dispatcher, $this->factory);
+    }
+
+    protected function tearDown()
+    {
+        if (in_array(ValidatorExtensionTrait::class, class_uses($this))) {
+            $this->validator = null;
+        }
+    }
+
+    protected function getExtensions()
+    {
+        $extensions = array();
+
+        if (in_array(ValidatorExtensionTrait::class, class_uses($this))) {
+            $extensions[] = $this->getValidatorExtension();
+        }
+
+        return $extensions;
     }
 
     public static function assertDateTimeEquals(\DateTime $expected, \DateTime $actual)
