+++ src/Symfony/Component/Form/Tests/Extension/Validator/Type/SubmitTypeValidatorExtensionTest.php
@@ -11,8 +11,12 @@
 
 namespace Symfony\Component\Form\Tests\Extension\Validator\Type;
 
+use Symfony\Component\Form\Test\Traits\ValidatorExtensionTrait;
+
 class SubmitTypeValidatorExtensionTest extends BaseValidatorExtensionTest
 {
+    use ValidatorExtensionTrait;
+
     protected function createForm(array $options = array())
     {
         return $this->factory->create('Symfony\Component\Form\Extension\Core\Type\SubmitType', null, $options);
